pyproject.toml
[tool.poetry]
name = "carbonmail"
version = "0.1.0"
description = ""
authors = ["Felipe Cabrera <felipe@bylearn.com.br>"]

[tool.poetry.dependencies]
python = "^3.8"
PySimpleGUI = "^4.38.0"
requests = "^2.25.1"
python-decouple = "^3.4"
pyinstaller = "^4.2"
pysimplegui-exemaker = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
